Sua 2 bang

+++ TravelezWeb/src/main/java/com/example/travelezweb/repository/TourRepository.java
@@ -2,7 +2,6 @@ package com.example.travelezweb.repository;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.travelezweb.model.Tour;
-public interface TourRepository  extends JpaRepository<Tour, Long> {
-
+public interface TourRepository  extends JpaRepository<Tour, Integer> {
 
 }
